+++ RELEASES.md
@@ -1,4 +1,8 @@
+# 0.7.1
 ## Added
+- `Mul` implementations for `{G1Affine, G1Projective, G2Affine, G2Projective}`
+  with `Scalar` on the left-hand side. These enable code to be written matching
+  equations that use the additive notation `[a] P`.
 - Zeroization support for `Gt`.
 
 # 0.7.0